requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-docx
langchain
openai
requests
python-multipart
pydantic
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
beautifulsoup4
PyJWT>=2.4.0
mammoth
natsort
pdfplumber
langchain-ollama
langchain-openai
docxcompose
htmldocx
httpx
langchain-core